Add comments to main thread modules
